Increment OpenMPI version (#15)
* Increment OpenMPI version.

* Comment.

* Whitespace.

* Whitespace.

* Increment the mpi4py version.

* Proxy on command line.

* Add a return.

* Formatting.

* Remove return.

* Use printf.

+++ tools/install.sh
@@ -2,6 +2,18 @@
 
 set -e
 
+# parse command line
+PROXY=""
+while [ "${1}" != "" ]; do
+    case ${1} in
+        --proxy )
+            shift
+            PROXY=${1}
+            ;;
+    esac
+    shift
+done
+
 # store location of this script
 TOOLS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
@@ -17,29 +29,25 @@ conda env remove --yes --name spotlight
 
 # create Anaconda env
 conda create --yes --name spotlight python=${PYTHON_VERSION}
+
+# install GSAS-II
+# must be installed from base environment
+conda install --yes --channel briantoby --name spotlight gsas2pkg=1.1
+
+# enter Anaconda env
 conda activate spotlight
 
-# install OpenMPI
-mkdir -p ${CONDA_PREFIX}/src && cd ${CONDA_PREFIX}/src
-wget https://www.open-mpi.org/software/ompi/v2.1/downloads/openmpi-2.1.3.tar.gz
-tar -xvf openmpi-2.1.3.tar.gz
-cd openmpi-2.1.3
-CFLAGS=-O3 \
-CXXFLAGS=-O3 \
-./configure --prefix=${CONDA_PREFIX}
-make -j $(getconf _NPROCESSORS_ONLN) install
-
-# install Subversion
-conda install --yes --channel anaconda svn==1.9.7
-
 # install GSAS which requires Python 2.7 for installation scripts
-conda install --yes python==2.7.16
+conda install --yes --channel anaconda svn==1.9.7
 mkdir -p ${CONDA_PREFIX}/gsas && cd ${CONDA_PREFIX}/gsas
 curl https://subversion.xray.aps.anl.gov/trac/EXPGUI/browser/gsas/linux/dist/bootstrap.py?format=txt > bootstrap.py
-echo $'proxyout.lanl.gov\n8080' >> proxy.txt
-python bootstrap.py < proxy.txt
-rm proxy.txt
-conda install --yes python==${PYTHON_VERSION}
+if [ ${PROXY} ]; then
+    printf $"${PROXY}" >> proxy.txt
+    python2 bootstrap.py < proxy.txt
+    rm proxy.txt
+else
+    python2 bootstrapy.py noproxy
+fi
 
 # install gsaslanguage scripts and fix some issues
 cd ${CONDA_PREFIX}/gsas
@@ -48,22 +56,24 @@ cd scripts
 git reset --hard fe73549
 chmod +x ${CONDA_PREFIX}/gsas/scripts/gsas_get_current_wtfrac_esd
 
+# install OpenMPI
+conda install --yes gxx_linux-64
+mkdir -p ${CONDA_PREFIX}/src && cd ${CONDA_PREFIX}/src
+wget https://www.open-mpi.org/software/ompi/v3.0/downloads/openmpi-3.0.6.tar.gz
+tar -xvf openmpi-3.0.6.tar.gz
+cd openmpi-3.0.6
+CFLAGS=-O3 \
+CXXFLAGS=-O3 \
+./configure --prefix=${CONDA_PREFIX}
+make -j $(getconf _NPROCESSORS_ONLN) install
+
 # install Python packages for Spotlight
 python -m pip install --upgrade pip
 python -m pip install --requirement ${TOOLS_DIR}/../requirements.txt
 
-# install GSAS-II
-# note proxy input is broken in GSAS-II bootstrap script
-#conda install --yes pyopengl==3.1.1a1 r-maps==3.1.0 wxpython==4.0.4
-conda install --yes pyopengl r-maps wxpython
+# construct GSAS-II
 conda install --yes scons==3.1.0
-python -m pip install scipy==1.3.1
-mkdir -p ${CONDA_PREFIX}/gsasii && cd ${CONDA_PREFIX}/gsasii
-curl https://subversion.xray.aps.anl.gov/pyGSAS/install/bootstrap.py > bootstrap.py
-echo $'proxyout.lanl.gov\n8080' >> proxy.txt
-python bootstrap.py < proxy.txt
-rm proxy.txt
-cd fsource
+cd ${CONDA_PREFIX}/GSASII/fsource
 scons
 
 # install required packages
@@ -71,7 +81,7 @@ conda install --yes pkg-config==0.29.2
 
 # install optional packages
 conda install --yes --channel conda-forge imagemagick
-pip install gprof2dot==2017.9.19
+python -m pip install gprof2dot==2017.9.19
 
 # install Spotlight
 cd ${TOOLS_DIR}/..
@@ -92,5 +102,5 @@ echo i > temp.txt
 mkdir -p ${CONDA_PREFIX}/etc/conda/activate.d
 echo export PGPLOT_FONT=\$\{CONDA_PREFIX\}/pgl/grfont.dat >> ${CONDA_PREFIX}/etc/conda/activate.d/post.sh
 echo export gsas=\$\{CONDA_PREFIX\}/gsas >> ${CONDA_PREFIX}/etc/conda/activate.d/post.sh
-echo export PYTHONPATH=\$\{PYTHONPATH\}:\$\{CONDA_PREFIX\}/gsasii:\$\{CONDA_PREFIX\}/gsasii/fsource >> ${CONDA_PREFIX}/etc/conda/activate.d/post.sh
+echo export PYTHONPATH=\$\{PYTHONPATH\}:\$\{CONDA_PREFIX\}/GSASII:\$\{CONDA_PREFIX\}/GSASII/fsource >> ${CONDA_PREFIX}/etc/conda/activate.d/post.sh
 echo export PATH=\$\{PATH\}:\$\{gsas\}/exe:\$\{gsas\}/scripts >> ${CONDA_PREFIX}/etc/conda/activate.d/post.sh
